qtdemux: prevent segfault after seeking with negative rate
After a seek, the position and segments info of qtdemux are
reset to null or invalid values. For reverse playback, qtdemux
will seek to a keyframe and play until the next keyframe, then
seek to another previous keyframe. This seeking function requires
the position and segments info of qtdemux to be properly initialized.

Due to a missing check for segment.rate in the movie demuxing loop
it would consider the current segment to have ended without having
it initialized, leading qtdemux to attempt to do a previous keyframe
seek without proper position and segment info data. This would cause
 a segfault. This patch fixes it by correctly checking for segment
boundaries taking the segment.rate into consideration.

https://bugzilla.gnome.org/show_bug.cgi?id=725104