curlhttpsink: allow for unconditionally disabling proxying
Previously if the proxy server hostname was the empty string
curlhttpsink would never even set the libcurl option. For libcurl
however, having a proxy server hostname be the empty string means that
proxying should be disabled even if environment variables might be set.
Now with the restriction lifted, doing this is allowed.

https://bugzilla.gnome.org/show_bug.cgi?id=728960