custom/glue: fix some file headers (licences and authors)
Most of the contents of these files come from their *.custom
counterparts of the 0.9.x version of the bindings, so I'm
reinstating their author/licence info here.

In some cases, the content is pristine to the previous version
(no modification since then), so for these cases the current
licence is replaced completely with the previous licence.

These were spotted while creating a first version of a debian
package (creating a debian/copyright info file requires you
to audit all this). This is not (yet) part of BGO#725937.